Organized assets and page components into folders.  Updated file paths.
