+++ src/components/NavBar.js
@@ -3,6 +3,8 @@ import { Link, useHistory } from "react-router-dom";
 import { logOut } from '../actions/user';
 import firebase from '../firebase';
 
+import Logo from '../assets/BasicGryphonLogo.png';
+
 const NavBar = () => {
   const isAuthenticated = useSelector(state => state.user.isAuthenticated );
   const dispatch = useDispatch();
@@ -22,7 +24,9 @@ const NavBar = () => {
   return (
     <nav className="navbar navbar-expand-lg navbar-light bg-light sticky-top float-right">
       <div className="container-fluid">
-        <Link to='/' className="navbar-brand">GryphHacks</Link>
+        <Link to='/' className="navbar-brand">
+          <img src={Logo} className="img-responsive" width="60px" alt="logo" style={{ marginLeft: '20px' }}/>
+        </Link>
         <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
           <span className="navbar-toggler-icon"></span>
         </button>
