+++ src/components/pages/FAQ.js
