+++ src/components/pages/NotFound.js
